refactor: move the RV64EMC -> PVM2 linker into nub
`javm-transpiler` was a 2191-line generic RISC-V linker that imported
javm-cap at exactly four lines, purely to emit a cap-shaped `Image`.
Section concatenation, AUIPC-pair resolution, ecall-marker rewriting,
fallthrough injection and PVM2 validation are ISA work: nub must be
able to do all of it without knowing JAVM exists.

- nub-linker: `elf.rs` moves verbatim (it never had cap awareness);
  `linker.rs` becomes `link.rs` with steps 1-6 unchanged and step 7
  emitting a `nub_program::ProgramBlob` instead of driving
  `ImageBuilder`. `TranspileError` becomes `LinkError`, gaining an
  `InvalidProgram` variant.
- javm-transpiler: 2191 -> 137 lines. `linker::link_elf` is now
  `image_from_blob(nub_linker::link_elf(elf)?)`; `layout.rs` keeps only
  the cnode-slot convention (65..68), which is the one genuinely JAVM
  part of the old module. Dead `tracing`/`tracing-test` deps dropped.
- javm-cap's `layout` becomes a re-export of `nub_program::abi`. Call
  sites spelling `javm_cap::layout::DATA_BASE` are untouched.

Two guards were dropped rather than ported: the hand-rolled
code-overlaps-DATA_BASE and data-past-4GiB checks are now
`ProgramBlob::new`'s job. They are exactly equivalent — the old code
rounded the code length up to a page before comparing against
MAX_CODE_SIZE, and since MAX_CODE_SIZE is page-aligned,
`round_up(len) <= MAX` iff `len <= MAX`.

`image_from_blob` must add regions in `Regions::iter()` order and the
code last: `ImageBuilder::build` packs its content-addressed arena in
insertion order, so a reorder would shift every emitted byte. Both the
ordering and the invariant are pinned by tests.

All 15 guest blobs remain byte-identical to a clean rebuild of the
parent commit, so every hash-tree-root and every pinned gas vector is
unchanged.
