+++ nub/nub-linker/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "nub-linker"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+description = "RISC-V ELF to PVM2 linker: emits a personality-free nub-program ProgramBlob"
+
+[dependencies]
+nub-program = { path = "../nub-program" }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+proptest = { workspace = true }
