+++ rust/javm-cap/src/layout.rs
@@ -1,18 +1,13 @@
 //! PVM2 guest virtual-address-space layout (ABI constants).
 //!
-//! These constants define where a transpiler-emitted Image's code and
-//! data regions map in the guest's 32-bit address space. They are part
-//! of the PVM2 ABI contract: the transpiler (`javm-transpiler`) bakes
-//! `PC = CODE_BASE + byte_offset` into endpoint entry PCs and native
-//! `auipc`/`jalr` resolution and lays data caps from [`DATA_BASE`] up,
-//! and every runtime (`nub-arch-x86`, `nub-arch-local`, `javm`) maps
-//! `Image.code` read-only at `CODE_BASE` and data at `DATA_BASE`.
+//! These are re-exports of [`nub_program::abi`], which owns them: they
+//! are PVM2 ISA/ABI facts, not capability facts, and every producer and
+//! consumer of a PVM2 program must agree on them whether or not a
+//! capability system is involved.
 //!
-//! The constants live here in `javm-cap` because it is the only crate
-//! every producer (transpiler) and consumer (each runtime) depends on.
-//! Code placement is a fixed protocol constant rather than an
-//! Image-supplied mapping entry: an untrusted Image must not get to
-//! choose where its code lands.
+//! They remain re-exported here because `javm_cap::layout::DATA_BASE`
+//! is the spelling used across the JAVM runtimes, and because an Image
+//! is laid out against exactly these constants:
 //!
 //! ```text
 //!   [0,         CODE_BASE)  unmapped — NULL guard (catch PC=0 / null deref)
@@ -20,22 +15,8 @@
 //!   [DATA_BASE, 4 GiB)      DATA     — stack / ro / rw / heap, RO|RW
 //! ```
 //!
-//! Code low (4 MiB) gives the null guard; data high (256 MiB) keeps the
-//! whole data region contiguous above code instead of wrapping around
-//! it. Both `[0, CODE_BASE)` and `[CODE_BASE + code, DATA_BASE)` are
-//! unmapped, so a stray fetch or load there faults.
+//! Code placement is a fixed protocol constant rather than an
+//! Image-supplied mapping entry: an untrusted Image must not get to
+//! choose where its code lands.
 
-/// Guest virtual address where the (single) code region maps read-only.
-/// A PVM PC is `CODE_BASE + byte_offset`. Sits at 4 MiB so `[0, 4 MiB)`
-/// is an unmapped null guard.
-pub const CODE_BASE: u32 = 0x0040_0000;
-
-/// Guest virtual address where the data region begins. All data caps
-/// (stack / ro / rw / heap) and instance overlays live in `[DATA_BASE,
-/// 4 GiB)`. At 256 MiB, well clear of the largest permitted code region.
-pub const DATA_BASE: u32 = 0x1000_0000;
-
-/// Maximum byte length of the code region. Code occupies `[CODE_BASE,
-/// CODE_BASE + code_len)` and must stay below `DATA_BASE`, so
-/// `code_len ≤ DATA_BASE − CODE_BASE` = 252 MiB.
-pub const MAX_CODE_SIZE: u32 = DATA_BASE - CODE_BASE;
+pub use nub_program::abi::{CODE_BASE, DATA_BASE, MAX_CODE_SIZE};
