+++ rust/javm-transpiler/tests/layout.rs
@@ -1,28 +1,35 @@
+//! The cnode-slot convention for transpiler-emitted Images.
+//!
+//! Region *geometry* is `nub_program::Regions`' business and is tested
+//! in `nub-program`; what this crate owns is the slot each region's
+//! `Cap::Data` is filed under.
+
 use javm_transpiler::layout::*;
+use nub_program::RegionKind;
 
 #[test]
-fn layout_minimal_stack_only() {
-    // Data is laid out from DATA_BASE upward.
-    let base = DATA_BASE / PVM_PAGE_SIZE;
-    let l = ProgramLayout::compute(1, 0, 0, 0);
-    assert_eq!(l.stack.cap_index, STACK_CAP_INDEX);
-    assert_eq!(l.stack.base_page, base);
-    assert_eq!(l.stack.page_count, 1);
-    assert!(l.ro.is_none());
-    assert!(l.rw.is_none());
-    assert!(l.heap.is_none());
-    assert_eq!(l.stack_top(), u64::from(DATA_BASE) + 4096);
-    assert_eq!(l.total_data_pages(), 1);
+fn each_region_kind_maps_to_its_conventional_slot() {
+    assert_eq!(cap_index(RegionKind::Stack), STACK_CAP_INDEX);
+    assert_eq!(cap_index(RegionKind::Ro), RO_CAP_INDEX);
+    assert_eq!(cap_index(RegionKind::Rw), RW_CAP_INDEX);
+    assert_eq!(cap_index(RegionKind::Heap), HEAP_CAP_INDEX);
+}
+
+/// Distinct slots, and above the low range a guest's own cnode uses.
+#[test]
+fn slots_are_distinct_and_in_the_reserved_range() {
+    let slots = [STACK_CAP_INDEX, RO_CAP_INDEX, RW_CAP_INDEX, HEAP_CAP_INDEX];
+    let mut sorted = slots.to_vec();
+    sorted.sort_unstable();
+    sorted.dedup();
+    assert_eq!(sorted.len(), slots.len(), "cap indices must be distinct");
+    assert!(slots.iter().all(|&s| s >= 64));
 }
 
 #[test]
-fn layout_full_stack_ro_rw_heap() {
-    let base = DATA_BASE / PVM_PAGE_SIZE;
-    let l = ProgramLayout::compute(2, 1, 1, 4);
-    assert_eq!(l.stack.base_page, base);
-    assert_eq!(l.ro.as_ref().unwrap().base_page, base + 2);
-    assert_eq!(l.rw.as_ref().unwrap().base_page, base + 3);
-    assert_eq!(l.heap.as_ref().unwrap().base_page, base + 4);
-    assert_eq!(l.stack_top(), u64::from(DATA_BASE) + 2 * 4096);
-    assert_eq!(l.total_data_pages(), 2 + 1 + 1 + 4);
+fn abi_constants_are_re_exported_from_nub_program() {
+    assert_eq!(CODE_BASE, nub_program::abi::CODE_BASE);
+    assert_eq!(DATA_BASE, nub_program::abi::DATA_BASE);
+    assert_eq!(MAX_CODE_SIZE, DATA_BASE - CODE_BASE);
+    assert_eq!(PVM_PAGE_SIZE, nub_program::abi::PAGE_SIZE);
 }
